training and segment GUI speedup
both trainingGui and editSegmetnsGui would save and refresh the UI each time a segment was edited, which takes up to 10 seconds for big datasets. I made them only save and update every so often.
For editSegmentsGui I had to create a global variable to keep track of the data instead of writing and reading it on the hard disk for each click.