small improvements to gateToolGui and editSegmentsGui
Changed the *.fig file, for gateToolGui it was messed up on my screen. for editSegmentsGui I moved the options form the top to the side so that I can enlarge the image further.
Changed a bug in gateToolGui for the case where handles.multi_clist had a single clist that was not a cell array variable.
suppressed some printing to the command line in superSeggerViewerGui